refactor email sending to new file #227
